Fix ImagePickerPropertyRow circular dependency [ci skip]
