+++ apps/src/applab/designMode.js
@@ -1,5 +1,4 @@
-/* global $, Dialog */
-// TODO (josh) - don't pass `Dialog` into `createModalDialog`.
+/* global $ */
 
 // TODO (brent) - make it so that we dont need to specify .jsx. This currently
 // works in our grunt build, but not in tests
@@ -8,7 +7,7 @@ var DesignModeBox = require('./DesignModeBox.jsx');
 var DesignModeHeaders = require('./DesignModeHeaders.jsx');
 var DesignProperties = require('./designProperties.jsx');
 var DesignToggleRow = require('./DesignToggleRow.jsx');
-var AssetManager = require('./assetManagement/AssetManager.jsx');
+var showAssetManager = require('./assetManagement/show.js');
 var elementLibrary = require('./designElements/library');
 var studioApp = require('../StudioApp').singleton;
 var _ = require('../utils').getLodash();
@@ -288,33 +287,6 @@ designMode.onClear = function() {
   document.getElementById('divApplab').innerHTML = Applab.levelHtml = "";
 };
 
-/**
- * Display the "Manage Assets" modal.
- * @param assetChosen {Function} Called when the user chooses an asset. The
- *   "Choose" button in the UI only appears if this optional param is provided.
- * @param typeFilter {String} The type of assets to show and allow to be
- *   uploaded.
- */
-designMode.showAssetManager = function(assetChosen, typeFilter) {
-  var codeDiv = document.createElement('div');
-  var showChoseImageButton = assetChosen && typeof assetChosen === 'function';
-  var dialog = studioApp.createModalDialog({
-    Dialog: Dialog,
-    contentDiv: codeDiv,
-    defaultBtnSelector: 'again-button',
-    id: 'manageAssetsModal'
-  });
-  React.render(React.createElement(AssetManager, {
-    typeFilter : typeFilter,
-    assetChosen: showChoseImageButton ? function (fileWithPath) {
-      dialog.hide();
-      assetChosen(fileWithPath);
-    } : null
-  }), codeDiv);
-
-  dialog.show();
-};
-
 function toggleDragging (enable) {
   var children = $('#divApplab').children();
   if (enable) {
@@ -433,7 +405,7 @@ designMode.configureDesignToggleRow = function () {
       screens: ['screen1'],
       onDesignModeButton: throttledDesignModeClick,
       onCodeModeButton: Applab.onCodeModeButton,
-      handleManageAssets: designMode.showAssetManager
+      handleManageAssets: showAssetManager
     }),
     designToggleRow
   );
